Add Deepseek chat modal to dashboard
